pyproject.toml
[project]
name = "common-rules-server"
version = "0.1.0"
description = "A Model Context Protocol server for providing common rules to AI IDEs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12.3",
    "pydantic>=2.7.4",
    "mcp>=1.9.3",   
    "pyyaml>=6.0.2",
]

[project.scripts]
common-rules = "common_rules_server.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
test = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "testcontainers>=4.10.0",
]

[tool.pytest.ini_options]
testpaths = ["src/test"]
python_files = ["test_*.py"]

[tool.hatch.build.targets.wheel]
packages = ["src/common_rules_server"]
